Control of errors added to updateProfileInfo, profileInfo and deleteUser

+++ backend/controllers/delete/deleteUser.js
@@ -1,16 +1,24 @@
-const db = require('../../db/delete')
+const db = require("../../db/delete");
+const { getUsersIdDB } = require("../../db/select");
 
-const deleteUser = async (req,res) => {
+const deleteUser = async (req, res) => {
+  const { id } = req.params;
 
-    const {id} = req.params
+  try {
+    const ids = await getUsersIdDB(id);
+    console.log(ids)
 
-    try {
-        result = await db.deleteUser(id)
-    } catch (e) {
-        res.send(e.message)
-        return
+    if (ids.indexOf(parseInt(id)) === -1) {
+      res.status(401).send("User not found");
+      return;
     }
-    res.send(result)
-}
 
-module.exports = deleteUser
+    result = await db.deleteUser(id);
+  } catch (e) {
+    res.send(e.message);
+    return;
+  }
+  res.send(result);
+};
+
+module.exports = deleteUser;
